refactor(runtime): move prop controllers to read-write state
