+++ .changeset/crisp-toys-pull.md
@@ -1,0 +1,5 @@
+---
+'@makeswift/runtime': patch
+---
+
+refactor: move prop controllers to read-write state
