+++ packages/runtime/src/components/builtin/Form/Form.tsx
@@ -33,7 +33,7 @@ import Button from '../Button'
 import type { TableColumn } from './types'
 import { Link } from '../../shared/Link'
 import { getBox, type BoxModelHandle } from '../../../state/modules/read-write/box-models'
-import { PropControllersHandle } from '../../../state/modules/prop-controller-handles'
+import { PropControllersHandle } from '../../../state/modules/read-write/prop-controllers'
 import { DescriptorsPropControllers } from '../../../prop-controllers/instances'
 import { useTableFormFieldRefs } from '../../hooks/useTableFormFieldRefs'
 import { cx } from '@emotion/css'
