+++ packages/runtime/src/runtimes/react/controls/rich-text/EditableText/editable-text.tsx
@@ -24,7 +24,7 @@ import { useBuilderEditMode } from '../../..'
 import { DescriptorsPropControllers } from '../../../../../prop-controllers/instances'
 import { withBlock, withTypography, withBuilder, onKeyDown } from '../../../../../slate'
 import { BuilderEditMode } from '../../../../../state/modules/builder-edit-mode'
-import { PropControllersHandle } from '../../../../../state/modules/prop-controller-handles'
+import { PropControllersHandle } from '../../../../../state/modules/read-write/prop-controllers'
 import { pollBoxModel } from '../../../poll-box-model'
 import { Element, Leaf } from '../components'
 import { Descriptors } from '../rich-text'
