+++ packages/runtime/src/runtimes/react/hooks/use-control-instances.ts
@@ -1,6 +1,7 @@
 import { ControlInstance } from '@makeswift/controls'
 
-import { getPropControllers } from '../../../state/read-only-state'
+import { getPropControllers } from '../../../state/read-write-state'
+import { isReadWriteState } from '../../../state/unified-state'
 import { useDocumentKey } from './use-document-context'
 import { useSelector } from './use-selector'
 
@@ -8,7 +9,7 @@ export function useControlInstances(elementKey: string): Record<string, ControlI
   const documentKey = useDocumentKey()
 
   return useSelector(state => {
-    if (documentKey == null) return null
+    if (documentKey == null || !isReadWriteState(state)) return null
 
     return getPropControllers(state, { documentKey, elementKey })
   })
