+++ packages/runtime/src/state/__tests__/react-builder-preview.test.ts
@@ -7,7 +7,7 @@ import { middlewareOptions } from '../toolkit'
 
 import { registerDocument, unregisterDocument } from '../shared-api'
 import { changeDocument } from '../host-api'
-import { registerComponentHandle } from '../actions/internal/read-only-actions'
+import { registerComponentHandle } from '../actions/internal/read-write-actions'
 
 import { createRootReducer } from '../read-write-state'
 import { readOnlyElementTreeMiddleware } from '../middleware/read-only-element-tree'
