+++ packages/runtime/src/state/actions/internal/read-only-action-types.ts
@@ -8,13 +8,6 @@ export const ReadOnlyActionTypes = {
 
   REGISTER_COMPONENT: 'REGISTER_COMPONENT',
   UNREGISTER_COMPONENT: 'UNREGISTER_COMPONENT',
-  REGISTER_COMPONENT_HANDLE: 'REGISTER_COMPONENT_HANDLE',
-  UNREGISTER_COMPONENT_HANDLE: 'UNREGISTER_COMPONENT_HANDLE',
-
-  REGISTER_PROP_CONTROLLERS: 'REGISTER_PROP_CONTROLLERS',
-  UNREGISTER_PROP_CONTROLLERS: 'UNREGISTER_PROP_CONTROLLERS',
-  REGISTER_PROP_CONTROLLERS_HANDLE: 'REGISTER_PROP_CONTROLLERS_HANDLE',
-  UNREGISTER_PROP_CONTROLLERS_HANDLE: 'UNREGISTER_PROP_CONTROLLERS_HANDLE',
 
   REGISTER_REACT_COMPONENT: 'REGISTER_REACT_COMPONENT',
   UNREGISTER_REACT_COMPONENT: 'UNREGISTER_REACT_COMPONENT',
