+++ packages/runtime/src/state/actions/internal/read-write-action-types.ts
@@ -1,5 +1,14 @@
 export const ReadWriteActionTypes = {
   CHANGE_ELEMENT_TREE: 'CHANGE_ELEMENT_TREE',
+
+  REGISTER_COMPONENT_HANDLE: 'REGISTER_COMPONENT_HANDLE',
+  UNREGISTER_COMPONENT_HANDLE: 'UNREGISTER_COMPONENT_HANDLE',
+
+  REGISTER_PROP_CONTROLLERS: 'REGISTER_PROP_CONTROLLERS',
+  UNREGISTER_PROP_CONTROLLERS: 'UNREGISTER_PROP_CONTROLLERS',
+  REGISTER_PROP_CONTROLLERS_HANDLE: 'REGISTER_PROP_CONTROLLERS_HANDLE',
+  UNREGISTER_PROP_CONTROLLERS_HANDLE: 'UNREGISTER_PROP_CONTROLLERS_HANDLE',
+
   REGISTER_MEASURABLE: 'REGISTER_MEASURABLE',
   UNREGISTER_MEASURABLE: 'UNREGISTER_MEASURABLE',
