+++ packages/runtime/src/state/modules/element-trees.ts
@@ -11,7 +11,7 @@ import { ReadOnlyActionTypes } from '../actions/internal/read-only-action-types'
 import { ReadWriteActionTypes } from '../actions/internal/read-write-action-types'
 
 import { getRootElement, type Document } from './read-only-documents'
-import { type DescriptorsByComponentType } from './prop-controllers'
+import { type DescriptorsByComponentType } from './prop-controller-descriptors'
 
 export type ElementTree = {
   elements: Map<string, Element>
