+++ packages/runtime/src/state/modules/read-write/prop-controllers.ts
@@ -1,7 +1,7 @@
 import { ControlInstance } from '@makeswift/controls'
-import { type Action, type UnknownAction, ActionTypes, isKnownAction } from '../actions'
-import type { DescriptorsPropControllers } from '../../prop-controllers/instances'
-import type { Descriptor } from '../../prop-controllers/descriptors'
+import { type Action, type UnknownAction, ActionTypes, isKnownAction } from '../../actions'
+import type { DescriptorsPropControllers } from '../../../prop-controllers/instances'
+import type { Descriptor } from '../../../prop-controllers/descriptors'
 
 export interface PropControllersHandle<
   T extends Record<string, Descriptor> = Record<string, Descriptor>,
