+++ packages/runtime/src/state/unified-state.ts
@@ -8,3 +8,5 @@ export type State = ReadOnlyState | ReadWriteState
 export type Dispatch = ThunkDispatch<State, unknown, Action>
 export type ReadOnlyReducer = Reducer<ReadOnlyState>
 export type ReadWriteDispatch = ThunkDispatch<ReadWriteState, unknown, Action>
+
+export const isReadWriteState = (state: State): state is ReadWriteState => 'boxModels' in state
